Add sort_by_field and sort_unstable_by_field to [T]
The method `[T]::sort_by_key()` must return a value, not a reference.
This works fine for sorting integers, but it is inconvenient to sort a
structure based on a string contained therein. The existing workaround
is to use `[T]::sort_by()` and run the comparison between two elements
manually, but this method is cleaner.

See #34162